Mon, Mar 11, 2024, 1:32 PM +05:30

+++ main.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
 
-int doubleIt(int number)
+short doubleIt(short number)
 {
     rerurn number*2;
 }
 
-int halveIt(int number)
+short halveIt(short number)
 {
     return number/2;
 }
